+++ Dockerfile
@@ -61,16 +61,15 @@ ENV TRINITY_HOME "/opt/trinityrnaseq-Trinity-v2.4.0/"
 ENV PATH "/opt/bowtie2-2.3.1:/opt/trinityrnaseq-Trinity-v2.4.0:/opt/FusionInspector-v1.0.1:/opt:$PATH"
 
 # Add wrapper scripts
-COPY star_fusion_pipeline.py /opt/star_fusion_pipeline.py
-COPY convert_star_to_bedpe.py /opt/convert_star_to_bedpe.py
-COPY gene-list /home/gene-list
-COPY save-list /home/save-list
-COPY delete-list /home/delete-list
-COPY test/star-fusion.fusion_candidates.final.abridged.FFPM /home/star-fusion.fusion_candidates.final.abridged.FFPM
-COPY test/FusionInspector.fusion_predictions.final.abridged.FFPM /home/FusionInspector.fusion_predictions.final.abridged.FFPM
+COPY pipeline /opt/pipeline
+COPY pipeline/data/gene-list /home/gene-list
+COPY pipeline/data/save-list /home/save-list
+COPY pipeline/data/delete-list /home/delete-list
+COPY pipeline/test/star-fusion.fusion_candidates.final.abridged.FFPM /home/star-fusion.fusion_candidates.final.abridged.FFPM
+COPY pipeline/test/FusionInspector.fusion_predictions.final.abridged.FFPM /home/FusionInspector.fusion_predictions.final.abridged.FFPM
 
 # Data processing occurs at /data
 WORKDIR /data
 
-ENTRYPOINT ["python", "/opt/star_fusion_pipeline.py"]
+ENTRYPOINT ["python", "/opt/pipeline/run.py"]
 CMD ["-h"]
