+++ pipeline/data/save-list
@@ -1,0 +1,9 @@
+Log.final.out$
+star-fusion.fusion_candidates.final.in_genelist.abridged$
+star-fusion.fusion_candidates.final.abridged.FFPM$
+FusionInspector.bed$
+FusionInspector.fa$
+FusionInspector.fusion_predictions.final.abridged.FFPM$
+FusionInspector.gmap_trinity_GG.fusions.fasta$
+FusionInspector.junction_reads.bam$
+FusionInspector.spanning_reads.bam$
