+++ pipeline/test/FusionInspector.fusion_predictions.final.abridged.FFPM
