+++ pipeline/test/star-fusion.fusion_candidates.final.abridged.FFPM
